+++ examples/components/vite/vite.config.ts
@@ -1,23 +1,16 @@
-import { resolve } from 'node:path';
-import { defineConfig } from 'vite';
-import { viteStaticCopy } from 'vite-plugin-static-copy';
+import { resolve } from "node:path";
+import { defineConfig, normalizePath } from "vite";
+import { viteStaticCopy } from "vite-plugin-static-copy";
 
 export default defineConfig({
   plugins: [
     viteStaticCopy({
       targets: [
         {
-          src: resolve(
-            'node_modules',
-            '@esri',
-            'calcite-components',
-            'dist',
-            'calcite',
-            'assets'
-          ),
-          dest: '.'
-        }
-      ]
-    })
-  ]
+          src: normalizePath(resolve("node_modules", "@esri", "calcite-components", "dist", "calcite", "assets")),
+          dest: normalizePath("."),
+        },
+      ],
+    }),
+  ],
 });